# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Perk Up Daily API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prompts:
    get:
      operationId: listPrompts
      tags: [prompts]
      summary: List reflection prompts
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReflectionPrompt"
    post:
      operationId: createPrompt
      tags: [prompts]
      summary: Create a reflection prompt (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReflectionPromptInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReflectionPrompt"
  /prompts/{id}:
    patch:
      operationId: updatePrompt
      tags: [prompts]
      summary: Update a reflection prompt (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReflectionPromptUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReflectionPrompt"
    delete:
      operationId: deletePrompt
      tags: [prompts]
      summary: Delete a reflection prompt (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /neurocycle:
    get:
      operationId: listNeurocycleCheckIns
      tags: [neurocycle]
      summary: List user neurocycle check-ins
      responses:
        "200":
          description: Check-ins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NeurocycleCheckIn"
    post:
      operationId: createNeurocycleCheckIn
      tags: [neurocycle]
      summary: Create a neurocycle check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NeurocycleCheckInInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NeurocycleCheckIn"
  /neurocycle/{id}:
    patch:
      operationId: updateNeurocycleCheckIn
      tags: [neurocycle]
      summary: Update a neurocycle check-in
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NeurocycleCheckInUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NeurocycleCheckIn"
...
  /users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: List all users (admin)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthUser"
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [auth]
      summary: Update a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
components:
  schemas:
...
    AuthUser:
      type: object
      required: [id, email]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        isPremium:
          type: boolean
        birthday:
          type: ["string", "null"]
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
...
    UserUpdate:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        birthday:
          type: ["string", "null"]
        onboardingCompleted:
          type: boolean
        isPremium:
          type: boolean
        dailyCheckInEnabled:
          type: boolean
        notificationsEnabled:
          type: boolean
...
    ReflectionPrompt:
      type: object
      required: [id, prompt, status]
      properties:
        id:
          type: integer
        prompt:
          type: string
        category:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    ReflectionPromptInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
        category:
          type: string
        status:
          type: string
    ReflectionPromptUpdate:
      type: object
      properties:
        prompt:
          type: string
        category:
          type: string
        status:
          type: string
...
    NeurocycleCheckIn:
      type: object
      required: [id, userId, day]
      properties:
        id:
          type: integer
        userId:
          type: integer
        day:
          type: integer
        stepData:
          type: ["object", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    NeurocycleCheckInInput:
      type: object
      required: [day]
      properties:
        day:
          type: integer
        stepData:
          type: object
        completedAt:
          type: string
    NeurocycleCheckInUpdate:
      type: object
      properties:
        day:
          type: integer
        stepData:
          type: object
        completedAt:
          type: string
...