# lib/api-zod/src/generated/types/neurocycleCheckIn.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Perk Up Daily API
 * OpenAPI spec version: 0.1.0
 */
import type { NeurocycleCheckInStepData } from './neurocycleCheckInStepData';

export interface NeurocycleCheckIn {
  id: number;
  userId: number;
  day: number;
  /** @nullable */
  stepData?: NeurocycleCheckInStepData;
  /** @nullable */
  completedAt?: string | null;
  createdAt?: string;
}