# lib/api-zod/src/generated/types/reflectionPromptUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Perk Up Daily API
 * OpenAPI spec version: 0.1.0
 */

export interface ReflectionPromptUpdate {
  prompt?: string;
  category?: string;
  status?: string;
}