# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Perk Up Daily API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  profilePicture?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  birthday?: string | null;
  onboardingCompleted?: boolean;
  isPremium?: boolean;
  dailyCheckInEnabled?: boolean;
  notificationsEnabled?: boolean;
}